clean up inference module boundaries
