+++ src/refiner/inference/__init__.py
@@ -1,6 +1,6 @@
 from refiner.inference.generate import generate
 from refiner.inference.generate_text import generate_text
-from refiner.inference.client import InferenceResponse
+from refiner.inference._response import InferenceResponse
 from refiner.inference.capabilities import ModelCapabilities, model_capabilities
 from refiner.inference._schema import InferenceSchemaValidationError
 from refiner.inference._transport import InferenceAPICallError, InferenceRetryError
