+++ src/refiner/inference/generate.py
@@ -4,7 +4,8 @@ from collections.abc import Awaitable, Callable, Mapping
 from typing import Any, TypeAlias
 
 from refiner.inference._runtime import inference_map
-from refiner.inference.client import InferenceResponse, _OpenAIEndpointClient
+from refiner.inference._response import InferenceResponse
+from refiner.inference.client import _OpenAIEndpointClient
 from refiner.inference.providers import OpenAIEndpointProvider, VLLMProvider
 from refiner.pipeline.data.row import Row
 from refiner.pipeline.steps import MapResult
